requirement.txt
alembic==1.2.1
cffi==1.13.1
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
